Add files via upload
A to-do list using JavaScript is a web application that allows users to create and manage a list of tasks they need to complete. It typically consists of a form where users can enter a task and a list of the tasks they have added, with the ability to mark tasks as complete or remove them from the list.